tests/wpt/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-remove-head.html

<!doctype html>
<title>CSS Color Adjust Test: Remove head with meta color-scheme</title>
<meta name="color-scheme" content="dark">
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<body></body>
<script>
  assert_root_color_scheme("dark", "Meta color-scheme applies.");
  document.head.remove();
  assert_root_color_scheme("normal", "Initial value after removing head including meta color-scheme.");
</script>
