LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-remove-head-expected.txt


FAIL Meta color-scheme applies. assert_equals: expected "dark" but got "normal"
PASS Initial value after removing head including meta color-scheme.
